Add GUID as cookie fordevice identification

+++ routes/login/confirm-device.js
@@ -1,13 +1,17 @@
 const { DB } = include("config/");
-const { deviceInfo, sendDeviceCode } = require("./device-info");
+const { deviceInfo, sendDeviceCode, checkDevice } = require("./device-info");
 const { updateRecord } = include("routes/save/contributor/confirm-device");
-
+const { v4: uuidv4 } = require('uuid');
 
 exports.confirmDevice = async (req, res, next) => {
   const { otp } = req.body;
   const { confirm_dev_origins } = req.session;
   const { redirecturl, uuid, u_profile } = confirm_dev_origins || {};
   const { sessionID: sid } = req || {};
+
+  const deviceGUID1 = uuidv4(); // Generate a unique GUID for the device
+  const deviceGUID2 = uuidv4();
+  const deviceGUID3 = uuidv4();
 
   req.session.errormessage = "";
   const device = deviceInfo(req);
@@ -29,9 +33,9 @@ exports.confirmDevice = async (req, res, next) => {
           // Code exists, add device info to the list of trusted devices
           return t.none(
             `
-                  INSERT INTO trusted_devices (user_uuid, device_name, device_os, device_browser, last_login, session_sid, is_trusted)
-                  VALUES ($1, $2, $3, $4, $5, $6, true)`,
-            [uuid, device.device, device.os, device.browser, new Date(), sid]
+                  INSERT INTO trusted_devices (user_uuid, device_name, device_os, device_browser, last_login, session_sid, duuid1, duuid2, duuid3, is_trusted)
+                  VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, true)`,
+            [uuid, device.device, device.os, device.browser, new Date(), sid, deviceGUID1, deviceGUID2, deviceGUID3]
           );
         } else {
           throw new Error("Invalid OTP");
@@ -62,6 +66,7 @@ exports.confirmDevice = async (req, res, next) => {
           req.session.destroy();
           res.redirect("/login");
         } else {
+          //SET USER SESSION EXPIRATION TO ONE YEAR
           const sessionExpiration = new Date(
             Date.now() + 365 * 24 * 60 * 60 * 1000
           ); // 1 year from now
@@ -73,11 +78,17 @@ exports.confirmDevice = async (req, res, next) => {
             ...device,
             is_trusted : true
           }
+
+          res.cookie('__ucd_app', deviceGUID1, { sessionExpiration }); 
+          res.cookie('__puid', deviceGUID2, { sessionExpiration }); 
+          res.cookie('__cduid', deviceGUID3, { sessionExpiration }); 
+          
           res.redirect(redirecturl);
           req.session.confirm_dev_origins = null;
         }
       })
       .catch((err) => {
+        console.log('err ', err)
         req.session.errormessage = "Invalid OTP";
         res.redirect("/confirm-device");
       });
@@ -108,10 +119,15 @@ exports.removeDevice = async (req, res) => {
 
   try {
     await DB.general.tx(async (t) => {
-      const sid = await t.oneOrNone('SELECT session_sid FROM trusted_devices WHERE id = $1 AND user_uuid = $2', [id, uuid], d => d.session_sid )
+      //ALLOW REMOVE OF DEVICE ONLY FROM TRUSTED DEVICES
+      const is_trusted = await checkDevice({req, conn: t})
+      if(is_trusted){
+        const sid = await t.oneOrNone('SELECT session_sid FROM trusted_devices WHERE id = $1 AND user_uuid = $2', [id, uuid], d => d.session_sid )
   
-      await t.none('DELETE FROM trusted_devices WHERE id = $1 AND user_uuid = $2', [id, uuid]);
-      await t.none('DELETE FROM session WHERE sid = $1', [sid]);
+        await t.none('DELETE FROM trusted_devices WHERE id = $1 AND user_uuid = $2', [id, uuid]);
+        await t.none('DELETE FROM session WHERE sid = $1', [sid]);
+      }
+
     });
     res.redirect(referer || `/${language}/edit/contributor?id=${uuid}`);
   } catch (err) {
